+++ admin/tabs/preferences/UserPreferencePanel.js
@@ -17,6 +17,7 @@ export class UserPreferencePanel extends Component {
     localModel = new RestGridModel({
         stateModel: 'xhUserPreferenceGrid',
         enableColChooser: true,
+        enableExport: true,
         store: new RestStore({
             url: 'rest/userPreferenceAdmin',
             reloadLookupsOnLoad: true,
