+++ columns/index.js
@@ -1,3 +1,4 @@
 export * from './Core';
 export * from './DatesTimes';
+export * from './ExportFormat';
 export * from './Utils';
